# Move global random number generator to `randRes` field