--- a/sdk/metric/exemplar_test.go
+++ b/sdk/metric/exemplar_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 )