@MrAlias 2024-09-13
Working to add a test that reproduces https://github.com/open-telemetry/opentelemetry-go/issues/5814