@MrAlias 2024-09-13
Verified this change passes the reproduction commented here: https://github.com/open-telemetry/opentelemetry-go/issues/5814#issuecomment-2349873146